move checkpointing to utils
